+++ src/types.ts
@@ -1,3 +1,5 @@
+import {ZeebeJob} from "zeebe-node";
+
 export interface InputVariables {
   to?: string | string[]
   cc?: string | string[]
@@ -14,3 +16,8 @@ export interface CustomHeaders {
     message: string
     pdfName?: string
 }
+
+// useful to make a readonly type to writeable, mainly for ZeebeJob
+type DeepWriteable<T> = { -readonly [P in keyof T]: DeepWriteable<T[P]> };
+
+export type ZeebeJobWriteable = DeepWriteable<ZeebeJob<InputVariables, CustomHeaders, OutputVariables>>