+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+*Undeployed*
+- Add the special hook to rewrite customHeaders
+
 *1.3.0* (2022-11-09)
 - Move test from a static ethereal account to a dynamically generated one